requirements.txt
streamlit
langgraph
langchain
langchain-openai
langchain-community
langchain-core
tavily-python
python-dotenv
duckduckgo-search